Add full BigTIFF read+write support, tests, and fuzzer
Reader fixes (A1-A3):
- Add TD_TYPE_SLONG8 case to td_val_uint_from_buf
- Add td_read_scalar_uint64 for uint64 output; widen JPEG_IF_OFFSET,
  JPEG_IF_BYTE_COUNT, EXIF_IFD read paths to uint64_t
- Widen jpeg_if_offset/byte_count in td_internal.h and jpeg_byte_count
  in tinydng.h to uint64_t

Writer (B1-B13):
- Add uint8_t bigtiff to tinydng_write_options
- Widen td_wentry.value to 8B, .count to uint64_t; widen
  td_writer_seg.byte_count to uint64_t
- Add td_put64 and td_add_long8 helpers
- Emit 16-byte header (v43) for BigTIFF
- Emit 20-byte IFD entries with 8-byte count/value, 8-byte next-pointer
- Use LONG8 (type 16) for strip/tile offset and byte-count arrays
- data_pos starts at 16 for BigTIFF; extras capacity 4MB
- Lift UINT32_MAX caps on IFD offset and segment size for BigTIFF

Tests:
- 4 BigTIFF roundtrip tests (none/LZW/LJPEG, mono/RGB, 8/16-bit)
- 5 BigTIFF decode tests: big-endian, tiled, LONG offsets, predictor2,
  SubIFDs
- All hand-built binary buffers with correct multi-byte endianness

Fuzzer:
- fuzz-v3-bigtiff.c: dual-path harness (reader fuzzing + BigTIFF
  write-read roundtrip with pixel-mismatch abort)
- 30 minutes, 1.39M executions, 1007 edges, zero crashes

Build: 7/7 CTest pass, clean build zero warnings.

+++ td_internal.h
@@ -271,8 +271,8 @@ typedef struct td_ifd_build {
   uint16_t predictor;
   uint32_t rows_per_strip;
   uint32_t tile_width, tile_length;
-  uint32_t jpeg_if_offset;
-  uint32_t jpeg_if_byte_count;
+  uint64_t jpeg_if_offset;
+  uint64_t jpeg_if_byte_count;
   uint32_t new_subfile_type;
 
   uint64_t *strip_offsets;
